+++ packages/moonstone/VirtualList/VirtualListBase.js
@@ -281,6 +281,7 @@ const VirtualListBaseFactory = (type) => {
 
 		isScrolledBy5way = false
 		isScrolledByJump = false
+		isWrappedBy5way = false
 		lastFocusedIndex = null
 		nodeIndexToBeFocused = null
 		preservedIndex = null
@@ -459,7 +460,7 @@ const VirtualListBaseFactory = (type) => {
 				focusedIndex = Number.parseInt(focusedItem.getAttribute(dataIndexAttribute)),
 				indexToScroll = this.getIndexToScroll(direction, focusedIndex);
 
-			if (indexToScroll !== -1) {
+			if (indexToScroll !== -1 && focusedIndex !== indexToScroll) {
 				const
 					isRtl = this.props.rtl,
 					isForward = (direction === 'down' || isRtl && direction === 'left' || !isRtl && direction === 'right');
@@ -479,9 +480,11 @@ const VirtualListBaseFactory = (type) => {
 					this.nodeIndexToBeFocused = this.lastFocusedIndex = indexToScroll;
 				}
 				cbScrollTo({index: indexToScroll, stickTo: isForward ? 'end' : 'start', animate: false});
-			}
 
-			return true;
+				return true;
+			} else {
+				return null;
+			}
 		}
 
 		/**
@@ -528,6 +531,9 @@ const VirtualListBaseFactory = (type) => {
 					!isPrimaryDirectionVertical && (!rtl && isLeft(keyCode) || rtl && isRight(keyCode)) ||
 					null
 				);
+
+			this.isScrolledBy5way = false;
+			this.isWrappedBy5way = false;
 
 			// If the currently focused item is disabled, we assume that all items in a list are disabled.
 			if (
@@ -603,6 +609,10 @@ const VirtualListBaseFactory = (type) => {
 			}
 
 			if (nextIndex !== -1) {
+				if (isWrapped) {
+					this.isWrappedBy5way = true;
+				}
+
 				if (firstIndex <= nextIndex && nextIndex < firstIndex + numOfItems) {
 					this.focusOnItem(nextIndex);
 				} else {
@@ -627,6 +637,7 @@ const VirtualListBaseFactory = (type) => {
 						}, 50);
 					}
 
+					this.isScrolledBy5way = true;
 					cbScrollTo({
 						index: nextIndex,
 						stickTo: isForward ? 'end' : 'start',
@@ -643,12 +654,10 @@ const VirtualListBaseFactory = (type) => {
 		onKeyDown = (ev) => {
 			const {keyCode, repeat, target} = ev;
 
-			this.isScrolledBy5way = false;
 			if (getDirection(keyCode)) {
 				ev.preventDefault();
 				this.setSpotlightContainerRestrict(keyCode, target);
-				this.isScrolledBy5way = this.jumpToSpottableItem(keyCode, repeat, target);
-				if (this.isScrolledBy5way) {
+				if (this.jumpToSpottableItem(keyCode, repeat, target)) {
 					ev.stopPropagation();
 				}
 			}
@@ -816,6 +825,8 @@ const VirtualListBaseFactory = (type) => {
 		}
 
 		shouldPreventScrollByFocus = () => ((type === JS) ? (this.isScrolledBy5way) : (this.isScrolledBy5way || this.isScrolledByJump))
+
+		shouldPreventOverscrollEffect = () => (this.isWrappedBy5way)
 
 		setLastFocusedIndex = (param) => {
 			this.lastFocusedIndex = param;