+++ packages/sampler/stories/moonstone-stories/Divider.js
@@ -1,12 +1,13 @@
 import Divider, {DividerBase} from '@enact/moonstone/Divider';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {text} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
-import {mergeComponentMetadata, smartSelect} from '../../src/utils';
+import {select, text} from '../../src/enact-knobs';
+import {mergeComponentMetadata} from '../../src/utils';
 
 const Config = mergeComponentMetadata('Divider', DividerBase, Divider);
+Divider.displayName = 'Divider';
 
 // Set up some defaults for info and knobs
 const prop = {
@@ -18,13 +19,16 @@ const prop = {
 storiesOf('Moonstone', module)
 	.add(
 		'Divider',
-		withInfo('Basic usage of divider')(() => (
+		withInfo({
+			propTablesExclude: [Divider],
+			text: 'Basic usage of Divider'
+		})(() => (
 			<Divider
-				casing={smartSelect('casing', prop.casing, Config)}
-				marqueeOn={smartSelect('marqueeOn', prop.marqueeOn, Config)}
-				spacing={smartSelect('spacing', prop.spacing, Config)}
+				casing={select('casing', prop.casing, Config)}
+				marqueeOn={select('marqueeOn', prop.marqueeOn, Config)}
+				spacing={select('spacing', prop.spacing, Config)}
 			>
-				{text('children', 'divider text')}
+				{text('children', Config, 'divider text')}
 			</Divider>
 		))
 	);