+++ packages/sampler/stories/moonstone-stories/ProgressBar.js
@@ -1,32 +1,34 @@
-import ProgressBar, {ProgressBarBase, ProgressBarTooltip} from '@enact/moonstone/ProgressBar';
+import ProgressBar, {ProgressBarTooltip} from '@enact/moonstone/ProgressBar';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {boolean, number, select} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
-import nullify from '../../src/utils/nullify.js';
-import {mergeComponentMetadata} from '../../src/utils/propTables';
+import {boolean, number, select} from '../../src/enact-knobs';
+import {mergeComponentMetadata} from '../../src/utils';
 
-const Config = mergeComponentMetadata('ProgressBar', ProgressBarBase, ProgressBar);
+const Config = mergeComponentMetadata('ProgressBar', ProgressBar);
+
+ProgressBar.displayName = 'ProgressBar';
+ProgressBarTooltip.displayName = 'ProgressBarTooltip';
 
 storiesOf('Moonstone', module)
 	.add(
 		'ProgressBar',
 		withInfo({
-			propTables: [Config],
+			propTablesExclude: [ProgressBar, ProgressBarTooltip],
 			text: 'The basic ProgressBar'
 		})(() => {
-			const side = nullify(select('side', ['after', 'before', 'left', 'right'], 'before'));
-			const tooltip = nullify(boolean('tooltip', false));
+			const side = select('side', ['after', 'before', 'left', 'right'], ProgressBarTooltip, 'before');
+			const tooltip = boolean('tooltip', ProgressBarTooltip);
 
 			return (
 				<ProgressBar
-					backgroundProgress={number('backgroundProgress', 0.5, {range: true, min: 0, max: 1, step: 0.01})}
-					disabled={nullify(boolean('disabled', false))}
-					highlighted={nullify(boolean('highlighted', false))}
-					orientation={select('orientation', ['horizontal', 'vertical'], 'horizontal')}
-					progress={number('progress', 0.4, {range: true, min: 0, max: 1, step: 0.01})}
-					side={nullify(select('side', ['after', 'before', 'left', 'right'], 'before'))}
+					backgroundProgress={number('backgroundProgress', Config, {range: true, min: 0, max: 1, step: 0.01}, 0.5)}
+					disabled={boolean('disabled', Config)}
+					highlighted={boolean('highlighted', Config)}
+					orientation={select('orientation', ['horizontal', 'vertical'], Config, 'horizontal')}
+					progress={number('progress', Config, {range: true, min: 0, max: 1, step: 0.01}, 0.4)}
+					side={select('side', ['after', 'before', 'left', 'right'], Config, 'before')}
 				>
 					{tooltip ? (
 						<ProgressBarTooltip