+++ packages/sampler/stories/moonstone-stories/SelectableItem.js
@@ -1,30 +1,28 @@
-import ToggleItem from '@enact/moonstone/ToggleItem';
 import SelectableItem from '@enact/moonstone/SelectableItem';
-import Item, {ItemBase} from '@enact/moonstone/Item';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
-import {boolean, text} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
-import nullify from '../../src/utils/nullify.js';
-import {mergeComponentMetadata} from '../../src/utils/propTables';
+import {boolean, text} from '../../src/enact-knobs';
+import {mergeComponentMetadata} from '../../src/utils';
 
-const Config = mergeComponentMetadata('SelectableItem', ItemBase, Item, ToggleItem, SelectableItem);
+const Config = mergeComponentMetadata('SelectableItem', SelectableItem);
+SelectableItem.displayName = 'SelectableItem';
 
 storiesOf('Moonstone', module)
 	.add(
 		'SelectableItem',
 		withInfo({
-			propTables: [Config],
+			propTablesExclude: [SelectableItem],
 			text: 'Basic usage of SelectableItem'
 		})(() => (
 			<SelectableItem
-				disabled={boolean('disabled', false)}
-				inline={nullify(boolean('inline', false))}
+				disabled={boolean('disabled', Config)}
+				inline={boolean('inline', Config)}
 				onToggle={action('onToggle')}
 			>
-				{text('children', 'Hello SelectableItem')}
+				{text('children', Config, 'Hello SelectableItem')}
 			</SelectableItem>
 		))
 	);