+++ packages/sampler/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact Sampler, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- Knobs to incorporate several useful features such as labeling the default value in the selection knob.
+
 ## [2.0.0-beta.7] - 2018-06-11
 
 No significant changes.